-debated adding a backpropegation step, decided against it as it would be cumbersome to implement. -current plan is to optimise network by replacing the flattened array with an aray of structs of arrays, but still not sure how to implement. -updated scoring function to bea little better (1 is ideal, 0 is terrible)

+++ kernels/evolutionKernel.cu
@@ -2,7 +2,7 @@
 #include <thrust/random.h>
 
 
-__global__ void evolutionKern(kernelArray<double> vect, kernelArray<int> params, uint32_t in, int device_offset){
+__global__ void evolutionKern(kernelArray<double> vect, kernelArray<int> params, uint32_t in, size_t device_offset){
     const int idx = blockIdx.x * blockDim.x + threadIdx.x;
     const int ind = params.array[10];
     int you, partner, child;
