+++ kernels/normalizeKernel.cu
@@ -1,6 +1,6 @@
 #include <kernelDefs.h>
 
-__global__ void normalizeKern(kernelArray<double> vect, kernelArray<int> params, double *avgFitness,  int device_offset){
+__global__ void normalizeKern(kernelArray<double> vect, kernelArray<int> params, double *avgFitness,  size_t device_offset){
     const int idx = blockIdx.x * blockDim.x + threadIdx.x; // for each thread is one individual
     const int fitnessval = params.array[19] + idx + device_offset;
     vect.array[fitnessval] = vect.array[fitnessval]/(*avgFitness);
